api/bin/requirements.txt
numpy
tqdm
pillow